Remove structure.py
