+++ atomica/version.py
@@ -4,5 +4,5 @@ Atomica version file.
 Standard location for module version number and date.
 """
 
-version = "1.0.11"
+version = "1.0.12"
 versiondate = "2018-10-26"
